# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CSVJob:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      failed_rows:
        type: integer
      filename:
        type: string
      id:
        type: string
      processed_rows:
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/domain.CSVJobStatus'
      total_rows:
        type: integer
      updated_at:
        type: string
    type: object
  domain.CSVJobStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - CSVJobStatusPending
    - CSVJobStatusProcessing
    - CSVJobStatusCompleted
    - CSVJobStatusFailed
  domain.CSVProcessingProgress:
    properties:
      failed_rows:
        type: integer
      job_id:
        type: string
      message:
        type: string
      processed_rows:
        type: integer
      status:
        $ref: '#/definitions/domain.CSVJobStatus'
      success_rate:
        type: number
      total_rows:
        type: integer
    type: object
  domain.CSVUploadResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/domain.CSVJob'
        type: array
      message:
        type: string
    type: object
...
  domain.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  domain.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/domain.PaginationInfo'
    type: object
  domain.PaginationInfo:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /csv/jobs:
    get:
      description: Get all CSV processing jobs for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.CSVJob'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user CSV jobs
      tags:
      - CSV
  /csv/jobs/{job_id}/progress:
    get:
      description: Get real-time progress of a CSV processing job
      parameters:
      - description: CSV Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CSVProcessingProgress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get CSV job progress
      tags:
      - CSV
  /csv/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple CSV files for concurrent processing
      parameters:
      - description: CSV files to upload
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CSVUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload and process CSV files
      tags:
      - CSV
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"